+++ CHANGELOG.md
@@ -12,9 +12,11 @@ All notable changes to the Pony compiler and standard library will be documented
 
 ### Added
 
+- `pony_chain` runtime function
 
 ### Changed
 
+- The `pony_sendv` and `pony_sendv_single` runtime functions now take a message chain instead of a single message
 - Forbid struct finalisers ([PR #2202](https://github.com/ponylang/ponyc/pull/2202))
 
 ## [0.18.1] - 2017-08-25